requirements.txt
scipy
numpy
matplotlib
seaborn
py-slippi
pillow